pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "sam-demo"
version = "0.1.0"
description = "CLI for running SAM3 text-prompted image segmentation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.13.0",
    "av>=14.0.0",
    "huggingface-hub>=0.35.0",
    "kernels>=0.12.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "requests>=2.33.1",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.4.0",
]

[project.scripts]
sam = "main:main"

[tool.hatch.build.targets.wheel]
include = ["main.py", "src/"]
